pushed ex02

+++ ex00/detectType.cpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   detectType.cpp                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jocelyn <jocelyn@student.42.fr>            +#+  +:+       +#+        */
+/*   By: jpiquet <jpiquet@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/18 08:31:37 by jocelyn           #+#    #+#             */
-/*   Updated: 2026/02/18 11:40:58 by jocelyn          ###   ########.fr       */
+/*   Updated: 2026/02/23 11:33:53 by jpiquet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,9 +25,6 @@ bool	isInt( std::string str )
 {
 	if(str.find('.') != str.npos)
 		return false;
-	// long i = std::atol(str.c_str());
-	// if (i > INT_MAX || i < INT_MIN)
-	// 	return false;
 	return true;
 }
 
@@ -43,7 +40,7 @@ bool isFloat( std::string str )
 bool isDouble( std::string str )
 {
 	float n = std::atof(str.c_str());
-	if(n > FLT_MAX || n < -FLT_MAX)
+	if(n > DBL_MAX || n < -DBL_MAX)
 		return false;
 	else
 		return true;
